refactor(orchestrator): D38 Phase 1 split chat_runner helpers
Extract message/seal, WebSocket send, and MCP/Plan/Question gate helpers
into dedicated modules with re-exports; drop unused Request schema
re-exports from chat_runner/main.
