+++ services/orchestrator/src/main.py
@@ -160,38 +160,6 @@ from src.chat_runner import (
     _run_sessions,
     _human_decision_locks,
     _human_decision_inflight,
-    StartRunRequest,
-    ValidateWorkflowRequest,
-    SaveUserWorkflowRequest,
-    WorkspaceRequest,
-    RepositoryGroupRequest,
-    RepositoryGroupUpdateRequest,
-    AgentsSaveRequest,
-    AgentPromptGenerateRequest,
-    ModelsConfigRequest,
-    OpenCodeZenListRequest,
-    ModelTestRequest,
-    CustomImageModelRequest,
-    CustomChatModelRequest,
-    CustomVideoModelRequest,
-    CustomModelUpdateRequest,
-    ToolConnectRequest,
-    ReassignRequest,
-    HumanDecisionRequest,
-    SessionCreateRequest,
-    SkillsMountRequest,
-    SkillsToggleRequest,
-    McpRegisterRequest,
-    McpServerIdRequest,
-    McpSaveConfigRequest,
-    CliActivateProviderRequest,
-    CliActivateModelRequest,
-    ThemePreferenceRequest,
-    LanguagePreferenceRequest,
-    PermissionModeRequest,
-    FontSizePreferenceRequest,
-    AvatarPreferenceRequest,
-    UserNamePreferenceRequest,
     _skills_registry_payload,
     _session_workspace_fields,
     _touch_session,
@@ -256,9 +224,6 @@ from src.chat_runner import (
     _notify_workspace_files_changed,
     _send_validation_result,
     _send_log_event,
-    ShellSnapshotUpdateRequest,
-    _workspace_http_error,
-    _build_agent_prompt_skeleton_fallback,
     _extract_llm_text,
     _async_handoff_summarization_task,
     ws_run,
